:sparkles: Add MnemonicTests
